pyproject.toml
[project]
name = "django-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "djangorestframework>=3.16.0",
    "factory-boy>=3.3.3",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.0",
]
